% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{post}
\alias{post}
\title{dataset: post selector}
\format{
An object of class \code{selector} (inherits from \code{R6}) of length 19.
}
\usage{
post
}
\description{
dataset: post selector
}
\keyword{datasets}
