% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{post_time}
\alias{post_time}
\title{dataset: date selector}
\format{
An object of class \code{selector} (inherits from \code{R6}) of length 19.
}
\usage{
post_time
}
\description{
dataset: date selector
}
\keyword{datasets}
