% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{post_title}
\alias{post_title}
\title{dataset: title selector}
\format{
An object of class \code{selector} (inherits from \code{R6}) of length 19.
}
\usage{
post_title
}
\description{
dataset: title selector
}
\keyword{datasets}
